Update WuFix-main.CMD
some user is not i the admin group. 
So when you change reg values as HKCU = hkey current user, the reg value is changed for the admin .

We now check for the SID value instead and use HKEY_USERS\%SID% to change reg value for the user. 

so even if users is not in admin group - it will successfully change the value for 
print key open snippet tool & storage sense