Update WuFix-main.CMD
this took longer then exptected, but it now works.

i pushed code before trying it out right, it didnt do anything more then throw error and not change the any reg value. 

but now it does, tested on a none admin user. 


why use a .txt document?  its the only way I can set a value and keeep it before it overides it when open the same dokument as admin. - if some1 have a better and cleaner solution im all ears.